Camera and photo inventory
almost there

When you press the R key, the camera view will open up on the screen. You can then use the spacebar to take a photo. The script will automatically save the photo to your persistent data path and display a preview of the photo in one of the inventory slots.

Pressing the I key will open up the inventory panel, which displays all the photos you have taken in the camera view. If you have taken more than 10 photos, the oldest photo will be automatically deleted to make space for the newest one. You can click on any of the photo icons in the inventory to view a larger version of the photo.

The UI and panels are fucked up right now but I'm getting to it